# apps/api/openapi.yml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /weapons/danger:
    get:
      operationId: WeaponsController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_0
        - weapons
  /weapons:
    get:
      operationId: WeaponsController_getAllWeapons
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
    post:
      operationId: WeaponsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWeaponDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
  /weapons/{id}:
    get:
      operationId: WeaponsController_getWeapon
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_0
    put:
      operationId: WeaponsController_update
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWeaponDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
    delete:
      operationId: WeaponsController_delete
      parameters:
        - name: id
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_0
info:
  title: DQ Weapons API
  description: The DQ Weapons API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateWeaponDto:
      type: object
      properties:
        name:
          type: string
        attackPower:
          type: number
        attribute:
          type: string
          enum:
            - sword
            - bow
            - spear
            - axe
            - dagger
      required:
        - name
        - attackPower
        - attribute
    UpdateWeaponDto:
      type: object
      properties:
        name:
          type: string
        attackPower:
          type: number
        attribute:
          type: string
          enum:
            - sword
            - bow
            - spear
            - axe
            - dagger